finalize testing stuff
So basically, story with this PR, is I wanted to add a new phase to TaskRun and it broke a million tests everywhere. These tests were long and sucked to read and write. This is my fault (and maybe claude's a little bit) - I did a big refactor, and restructured the tests grouped by phase transition, which forces them all to be quite short and focused - no test ever does multiple Reconcile calls - this is by design